# [CIS-1458] Speed up bootstrap action