--- a/.github/actions/set-build-image-var/action.yml
+++ b/.github/actions/set-build-image-var/action.yml
@@ -3,5 +3,5 @@ description: 'Sets $ImageVersion with buil image version'
 runs:
   using: "composite"
   steps:
-    - run: echo "ImageVersion=$ImageVersion"  >> $GITHUB_ENV 
+    - run: echo "ImageVersion=$ImageVersion" >> $GITHUB_ENV
       shell: bash