--- a/.github/workflows/publish-release.yml
+++ b/.github/workflows/publish-release.yml
@@ -18,22 +18,7 @@ jobs:
         with:
           ssh-private-key: ${{ secrets.BOT_SSH_PRIVATE_KEY }}
       - uses: actions/checkout@v2
-      - name: "Set build image var"
-        run: echo "ImageVersion=$ImageVersion"  >> $GITHUB_ENV
-      - name: Cache RubyGems
-        uses: actions/cache@v2
-        id: rubygem-cache
-        with:
-          path: vendor/bundle
-          key: ${{ runner.os }}-${{ env.ImageVersion }}-gem-${{ hashFiles('**/Gemfile.lock') }}
-          restore-keys: ${{ runner.os }}-${{ env.ImageVersion }}-gem-
-      - name: Cache Mint
-        uses: actions/cache@v2
-        id: mint-cache
-        with:
-          path: /usr/local/lib/mint
-          key: ${{ runner.os }}-mint-${{ hashFiles('./Mintfile') }}
-          restore-keys: ${{ runner.os }}-mint-
+      - uses: ./.github/actions/set-build-image-var
       - name: Extract version from branch name (for release branches)
         if: startsWith(github.event.pull_request.head.ref, 'release/')
         run: |