### 🔗 Issue Links

- https://stream-io.atlassian.net/browse/CIS-1458

### 🎯 Goal

To speed up bootstrap action

### 📝 Summary

- Change `mint` cache path
- Update `mint` cache key
- Move the `bootstrap` cache to the action file
- Move `allure` installation commands to `bootstrap`
- Add `set-build-image-var` action where possible
- Provide a new token for Allure Github requests to avoid rate limit issue
- Add support for Apple Silicon brew directory in `SwiftLint Build Phase`

### 🎨 Showcase

Bootstrap time reduced from ≈11min to ≈40sec 🏎

| Before | After |
| ------ | ----- |
| ![before](https://user-images.githubusercontent.com/20794991/167107341-cdbc33b5-0204-475d-8fb0-38494df73b01.png)|![after](https://user-images.githubusercontent.com/20794991/167107380-18476b99-58a3-48ed-a91c-2a3ebc6256ae.png)|

### ☑️ Contributor Checklist
- [x] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [ ] Changelog is updated with client-facing changes
- [ ] New code is covered by unit tests
- [x] This change follows zero ⚠️ policy (required)
- [ ] Comparison screenshots added for visual changes
- [ ] Affected documentation updated (docusaurus, tutorial, CMS)

### 🎁 Meme

![gif](https://media.giphy.com/media/129s9ZDHNVjqmI/giphy.gif)